+++ plugins/tools.py
@@ -16,7 +16,7 @@
 • `{i}bots`
     Shows the number of bots in the current chat with their perma-link.
 
-• `{i}hl <a link>`
+• `{i}hl <a link> <text-optional>`
     Embeds the link with a whitespace as message.
 
 • `{i}id`
@@ -44,7 +44,7 @@ from asyncio.exceptions import TimeoutError as AsyncTimeout
 import cv2
 from google_trans_new import google_translator
 from htmlwebshot import WebShot
-from pyUltroid.functions.tools import metadata
+from pyUltroid.functions.tools import TgConverter, metadata
 from telethon.errors.rpcerrorlist import MessageTooLongError, YouBlockedUserError
 from telethon.tl.types import (
     ChannelParticipantAdmin,
@@ -88,11 +88,11 @@ async def _(event):
 
 
 @ultroid_cmd(
-    pattern="id ?(.*)",
+    pattern="id( (.*)|$)",
     manager=True,
 )
 async def _(event):
-    match = event.pattern_match.group(1)
+    match = event.pattern_match.group(1).strip()
     if event.reply_to_msg_id:
         await event.get_input_chat()
         r_msg = await event.get_reply_message()
@@ -131,10 +131,10 @@ async def _(event):
         )
 
 
-@ultroid_cmd(pattern="bots ?(.*)", groups_only=True, manager=True)
+@ultroid_cmd(pattern="bots( (.*)|$)", groups_only=True, manager=True)
 async def _(ult):
     mentions = "• **Bots in this Chat**: \n"
-    input_str = ult.pattern_match.group(1)
+    input_str = ult.pattern_match.group(1).strip()
     if not input_str:
         chat = ult.chat_id
     else:
@@ -158,14 +158,20 @@ async def _(ult):
 
 
 @ultroid_cmd(
-    pattern="hl",
+    pattern="hl( (.*)|$)",
 )
 async def _(ult):
-    try:
-        input = ult.text.split(" ", maxsplit=1)[1]
-    except IndexError:
+    input_ = ult.pattern_match.group(1).strip()
+    if not input_:
         return await ult.eor("`Input some link`", time=5)
-    await ult.eor("[ㅤㅤㅤㅤㅤㅤㅤ](" + input + ")", link_preview=False)
+    text = None
+    if len(input_.split()) > 1:
+        spli_ = input_.split()
+        input_ = spli_[0]
+        text = spli_[1]
+    if not text:
+        text = "ㅤㅤㅤㅤㅤㅤㅤ"
+    await ult.eor(f"[{text}](" + input_ + ")", link_preview=False)
 
 
 @ultroid_cmd(
@@ -192,8 +198,8 @@ async def _(e):
             f'ffmpeg -i "{thumb}" -i "{audio.name}" -preset ultrafast -c:a libmp3lame -ab 64 circle.mp4 -y'
         )
         await msg.edit("`Uploading...`")
-        file, _ = await e.client.fast_uploader("circle.mp4", to_delete=True)
         data = await metadata("circle.mp4")
+        file, _ = await e.client.fast_uploader("circle.mp4", to_delete=True)
         await e.client.send_file(
             e.chat_id,
             file,
@@ -213,24 +219,30 @@ async def _(e):
     elif mediainfo(reply.media) == "gif" or mediainfo(reply.media).startswith("video"):
         msg = await e.eor("**Cʀᴇᴀᴛɪɴɢ Vɪᴅᴇᴏ Nᴏᴛᴇ**")
         file = await reply.download_media("resources/downloads/")
-        await e.client.send_file(
-            e.chat_id,
-            file,
-            video_note=True,
-            thumb="resources/extras/ultroid.jpg",
-            reply_to=reply,
-        )
+        if file.endswith(".webm"):
+            nfile = await TgConverter.ffmpeg_convert(file, "file.mp4")
+            os.remove(file)
+            file = nfile
+        if file:
+            await e.client.send_file(
+                e.chat_id,
+                file,
+                video_note=True,
+                thumb="resources/extras/ultroid.jpg",
+                reply_to=reply,
+            )
+            os.remove(file)
         await msg.delete()
-        os.remove(file)
+
     else:
         await e.eor("`Reply to a gif or audio file only.`")
 
 
 @ultroid_cmd(
-    pattern="ls ?(.*)",
+    pattern="ls( (.*)|$)",
 )
 async def _(e):
-    files = e.pattern_match.group(1)
+    files = e.pattern_match.group(1).strip()
     if not files:
         files = "*"
     elif files.endswith("/"):
@@ -350,10 +362,10 @@ async def _(e):
 
 
 @ultroid_cmd(
-    pattern="sg ?(.*)",
+    pattern="sg( (.*)|$)",
 )
 async def lastname(steal):
-    mat = steal.pattern_match.group(1)
+    mat = steal.pattern_match.group(1).strip()
     if not steal.is_reply and not mat:
         return await steal.eor("`Use this command with reply or give Username/id...`")
     if mat:
@@ -400,10 +412,10 @@ async def lastname(steal):
         await lol.edit("Error: @SangMataInfo_bot is not responding!.")
 
 
-@ultroid_cmd(pattern="webshot ?(.*)")
+@ultroid_cmd(pattern="webshot( (.*)|$)")
 async def webss(event):
     xx = await event.eor(get_string("com_1"))
-    xurl = event.pattern_match.group(1)
+    xurl = event.pattern_match.group(1).strip()
     if not xurl:
         return await xx.eor(get_string("wbs_1"), time=5)
     if not is_url_ok(xurl):
@@ -440,7 +452,7 @@ async def magic(event):
     )
     response = data.get("response", {})
     if not response.get("status"):
-        return await event.eor("**ERRROR :** `{}`".format(response["message"]))
+        return await event.eor("**ERROR :** `{}`".format(response["message"]))
     await event.eor(
         f"• **Ultroid Tiny**\n• Given Url : {url}\n• Shorten Url : {data['response']['tinyUrl']}"
     )
