+++ plugins/utilities.py
@@ -66,6 +66,7 @@ from pygments.lexers import Python3Lexer
 from pyUltroid.dB.gban_mute_db import is_gbanned
 from pyUltroid.misc._assistant import asst_cmd
 from telegraph import upload_file as uf
+from telethon.errors.rpcerrorlist import UserBotError
 from telethon.events import NewMessage
 from telethon.tl.custom import Dialog
 from telethon.tl.functions.channels import (
@@ -95,6 +96,7 @@ from . import (
     get_paste,
     get_string,
     inline_mention,
+    json_parser,
     mediainfo,
     udB,
     ultroid_bot,
@@ -131,11 +133,11 @@ async def date(event):
 )
 async def _(event):
     result = await event.client(GetAdminedPublicChannelsRequest())
-    r = result.chats
-    if not r:
+    if not result.chats:
         return await event.eor("`No username Reserved`")
     output_str = "".join(
-        f"- {channel_obj.title} @{channel_obj.username} \n" for channel_obj in r
+        f"- {channel_obj.title} @{channel_obj.username} \n"
+        for channel_obj in result.chats
     )
     await event.eor(output_str)
 
@@ -215,7 +217,7 @@ async def stats(
     await ok.edit(response)
 
 
-@ultroid_cmd(pattern="paste ?(.*)", manager=True, allow_all=True)
+@ultroid_cmd(pattern="paste( (.*)|$)", manager=True, allow_all=True)
 async def _(event):
     try:
         input_str = event.text.split(maxsplit=1)[1]
@@ -263,11 +265,11 @@ async def _(event):
 
 
 @ultroid_cmd(
-    pattern="info ?(.*)",
+    pattern="info( (.*)|$)",
     manager=True,
 )
 async def _(event):
-    match = event.pattern_match.group(1)
+    match = event.pattern_match.group(1).strip()
     if match:
         try:
             user = await event.client.parse_id(match)
@@ -292,7 +294,7 @@ async def _(event):
             if not photo:
                 return await xx.eor(capt, parse_mode="html")
             await event.client.send_message(
-                event.chat_id, capt, file=photo, parse_mode="html"
+                event.chat_id, capt[:1024], file=photo, parse_mode="html"
             )
             await xx.delete()
         except Exception as er:
@@ -322,7 +324,7 @@ async def _(event):
         dc_id = user.photo.dc_id
     else:
         dc_id = "Need a Profile Picture to check this"
-    caption = """<b>Exᴛʀᴀᴄᴛᴇᴅ Dᴀᴛᴀʙᴀsᴇ Fʀᴏᴍ Tᴇʟᴇɢʀᴀᴍ's Dᴀᴛᴀʙᴀsᴇ<b>
+    caption = """<b>Exᴛʀᴀᴄᴛᴇᴅ Dᴀᴛᴀ Fʀᴏᴍ Tᴇʟᴇɢʀᴀᴍ's Dᴀᴛᴀʙᴀsᴇ<b>
 <b>••Tᴇʟᴇɢʀᴀᴍ ID</b>: <code>{}</code>
 <b>••Pᴇʀᴍᴀɴᴇɴᴛ Lɪɴᴋ</b>: <a href='tg://user?id={}'>Click Here</a>
 <b>••Fɪʀsᴛ Nᴀᴍᴇ</b>: <code>{}</code>
@@ -366,12 +368,12 @@ async def _(event):
 
 
 @ultroid_cmd(
-    pattern="invite ?(.*)",
+    pattern="invite( (.*)|$)",
     groups_only=True,
 )
 async def _(ult):
     xx = await ult.eor(get_string("com_1"))
-    to_add_users = ult.pattern_match.group(1)
+    to_add_users = ult.pattern_match.group(1).strip()
     if not ult.is_channel and ult.is_group:
         for user_id in to_add_users.split(" "):
             try:
@@ -395,30 +397,40 @@ async def _(ult):
                     ),
                 )
                 await xx.edit(f"Successfully invited `{user_id}` to `{ult.chat_id}`")
+            except UserBotError:
+                await xx.edit(
+                    f"Bots can only be added as Admins in Channel.\nBetter Use `{HNDLR}promote {user_id}`"
+                )
             except Exception as e:
                 await xx.edit(str(e))
 
 
 @ultroid_cmd(
-    pattern=r"rmbg$",
+    pattern="rmbg($| (.*))",
 )
 async def abs_rmbg(event):
     RMBG_API = udB.get_key("RMBG_API")
     if not RMBG_API:
-        return await eor(
-            event,
+        return await event.eor(
             "Get your API key from [here](https://www.remove.bg/) for this plugin to work.",
         )
-    if not event.reply_to_msg_id:
+    match = event.pattern_match.group(1).strip()
+    reply = await event.get_reply_message()
+    if match and os.path.exists(match):
+        dl = match
+    elif reply and reply.media:
+        if reply.document and reply.document.thumbs:
+            dl = await reply.download_media(thumb=-1)
+        else:
+            dl = await reply.download_media()
+    else:
         return await eod(
             event, f"Use `{HNDLR}rmbg` as reply to a pic to remove its background."
         )
-    reply = await event.get_reply_message()
-    dl = await event.client.download_media(reply.media)
-    if not dl.endswith(("webp", "jpg", "png", "jpeg")):
+    if not (dl and dl.endswith(("webp", "jpg", "png", "jpeg"))):
         os.remove(dl)
         return await event.eor(get_string("com_4"))
-    if mediainfo(reply.media) == "sticker":
+    if dl.endswith("webp"):
         file = dl + ".png"
         Image.open(dl).save(file)
         os.remove(dl)
@@ -428,7 +440,7 @@ async def abs_rmbg(event):
     os.remove(dl)
     if not dn:
         dr = out["errors"][0]
-        de = dr["detail"] if dr.get("detail") else ""
+        de = dr.get("detail", "")
         return await xx.edit(
             f"**ERROR ~** `{dr['title']}`,\n`{de}`",
         )
@@ -450,10 +462,10 @@ async def abs_rmbg(event):
 
 
 @ultroid_cmd(
-    pattern="telegraph ?(.*)",
+    pattern="telegraph( (.*)|$)",
 )
 async def telegraphcmd(event):
-    match = event.pattern_match.group(1) or "Ultroid"
+    match = event.pattern_match.group(1).strip() or "Ultroid"
     reply = await event.get_reply_message()
     if not reply:
         return await event.eor("`Reply to Message.`")
@@ -489,19 +501,28 @@ async def telegraphcmd(event):
     )
 
 
-@ultroid_cmd(pattern="json$")
+@ultroid_cmd(pattern="json( (.*)|$)")
 async def _(event):
-    the_real_message = None
     reply_to_id = None
+    match = event.pattern_match.group(1).strip()
     if event.reply_to_msg_id:
-        previous_message = await event.get_reply_message()
-        the_real_message = previous_message.stringify()
+        msg = await event.get_reply_message()
         reply_to_id = event.reply_to_msg_id
     else:
-        the_real_message = event.stringify()
+        msg = event
         reply_to_id = event.message.id
-    if len(the_real_message) > 4096:
-        with io.BytesIO(str.encode(the_real_message)) as out_file:
+    if match and hasattr(msg, match):
+        msg = getattr(msg, match)
+        if hasattr(msg, "to_json"):
+            try:
+                msg = json_parser(msg.to_json(), indent=1)
+            except Exception as e:
+                LOGS.exception(e)
+        msg = str(msg)
+    else:
+        msg = json_parser(msg.to_json(), indent=1)
+    if len(msg) > 4096:
+        with io.BytesIO(str.encode(msg)) as out_file:
             out_file.name = "json-ult.txt"
             await event.client.send_file(
                 event.chat_id,
@@ -512,10 +533,10 @@ async def _(event):
             )
             await event.delete()
     else:
-        await event.eor(f"```{the_real_message}```")
+        await event.eor(f"```{msg or None}```")
 
 
-@ultroid_cmd(pattern="suggest ?(.*)", manager=True)
+@ultroid_cmd(pattern="suggest( (.*)|$)", manager=True)
 async def sugg(event):
     sll = event.text.split(maxsplit=1)
     try:
@@ -547,7 +568,7 @@ async def sugg(event):
     await event.delete()
 
 
-@ultroid_cmd(pattern="ipinfo ?(.*)")
+@ultroid_cmd(pattern="ipinfo( (.*)|$)")
 async def ipinfo(event):
     ip = event.text.split()
     ipaddr = ""
