+++ plugins/webupload.py
@@ -4,6 +4,7 @@
 # This file is a part of < https://github.com/TeamUltroid/Ultroid/ >
 # PLease read the GNU Affero General Public License in
 # <https://www.github.com/TeamUltroid/Ultroid/blob/main/LICENSE/>.
+
 """
 ✘ Commands Available -
 
@@ -19,11 +20,11 @@ from . import asst, get_string, ultroid_cmd
 
 
 @ultroid_cmd(
-    pattern="webupload ?(.*)",
+    pattern="webupload( (.*)|$)",
 )
 async def _(event):
     xx = await event.eor(get_string("com_1"))
-    match = event.pattern_match.group(1)
+    match = event.pattern_match.group(1).strip()
     if event.chat_id not in _webupload_cache:
         _webupload_cache.update({int(event.chat_id): {}})
     if match:
@@ -33,7 +34,7 @@ async def _(event):
     elif event.reply_to_msg_id:
         reply = await event.get_reply_message()
         if reply.photo:
-            file = await event.client.download_media("resources/downloads/")
+            file = await reply.download_media("resources/downloads/")
             _webupload_cache[int(event.chat_id)][int(event.id)] = file
         else:
             file, _ = await event.client.fast_downloader(
@@ -42,8 +43,27 @@ async def _(event):
             _webupload_cache[int(event.chat_id)][int(event.id)] = file.name
     else:
         return await xx.eor("`Reply to file or give file path...`")
-    results = await event.client.inline_query(
-        asst.me.username, f"fl2lnk {event.chat_id}:{event.id}"
-    )
-    await results[0].click(event.chat_id, reply_to=event.reply_to_msg_id)
-    await xx.delete()
+    if not event.client._bot:
+        results = await event.client.inline_query(
+            asst.me.username, f"fl2lnk {event.chat_id}:{event.id}"
+        )
+        await results[0].click(event.chat_id, reply_to=event.reply_to_msg_id)
+        await xx.delete()
+
+    else:
+        __cache = f"{event.chat_id}:{event.id}"
+        buttons = [
+            [
+                Button.inline("anonfiles", data=f"flanonfiles//{__cache}"),
+                Button.inline("transfer", data=f"fltransfer//{__cache}"),
+            ],
+            [
+                Button.inline("bayfiles", data=f"flbayfiles//{__cache}"),
+                Button.inline("x0.at", data=f"flx0.at//{__cache}"),
+            ],
+            [
+                Button.inline("file.io", data=f"flfile.io//{__cache}"),
+                Button.inline("siasky", data=f"flsiasky//{__cache}"),
+            ],
+        ]
+        await xx.edit("**Choose Server to Upload File...**", buttons=buttons)
